1. "플레이어 공격 -> 적 대미지" 구현 TODO : "적 공격 -> 플레이어 대미지" + 피격&사망 모션 연결

+++ Assets/Tower/Scripts/Enemy/Data/EnemyBaseSO.cs
@@ -11,7 +11,7 @@ namespace Tower.Enemy.Data
         public float def;
 
         public float maxHp;
-        public float maxGp;
+        public int maxGp;
     }
 
 }
